fixes #74. The sum() method returns an instance of numpy.int64 if the allocation only contains integers.
An instance of numpy.int64 is not recognized as instance of int, as they are two different classes.

This causes the error message since the total investment is not recognized as integer.

Courious: trying with allocations equal to 20^(200), the returned number is an instance of "int" instead of "numpy.int64".