Feature/beta parameter (#84)
Adding computation of the beta parameter for the portfolio. In the
Capital Asset Pricing Model (CAPM), the beta parameter defines how risky
the portfolio is compared to the market.

The additional feature is written so that all the previous
codes and tests should still work: suppose the argument "market_index"
is not explicitly passed to the "build_portfolio" function. In that
case, the Market class instance is not created, and the beta parameter
of the portfolio is not computed or printed in properties.