New feature: Value at Risk (VaR) (#106)
The Value at Risk (VaR) is a way to evaluate the risk of a portfolio. It
measures the potential loss that a portfolio can have over a certain
period with a specified confidence level. With respect to the beta
parameter, variance, etc., VaR uses a probability distribution, and it
assumes that returns are normally distributed.

Using the variance-covariance method, we want to compute the portfolio's
value at risk (VaR) over the selected period (the 'freq' variable). The
Monte Carlo approach is left for further development.

---------

Co-authored-by: Frank Milthaler <fmilthaler@users.noreply.github.com>