Chore/auto versioning workflow (#111)
This adds another GitHub workflow `AutoVersion` for automated
versioning.

`AutoVersion` is triggered after PR is merged into `master` or
`develop`. It then determines whether the version is a patch or minor
change based on the source branch name.
- `patch`: version number increase of `0.0.1`
- `minor`: version number increase of `0.1.0`
- `major`: version number increase of `1.0.0`

The branch name patterns that are currently matched to one of the above
are:
- `patch`: one of `chore/`, `refactor/`, `bugfix/`
- `minor`: `feature/`
- `major`: None (not planned at the moment)

Note: the patterns mentioned above, e.g. `bugfix/` must be the prefix of
the branch name. If you are working on a bugfix to fix a print statement
of the portfolio properties, your branch name should be something like
`bugfix/print-statement-portfolio-properties`. For the automated
versioning to work, the branch name is required to **start with**
`bugfix/` or one of the other above mentioned patterns.

For now, the release number is also updated every time the version
number is updated. This practically renders the release number
redundant, but it is used in the README.md to show the current release
number on PyPI.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>