Introducing the R-squared coefficient and Treynor Ratio for a financial portfolio (#134)
The R-squared coefficient measures how closely the portfolio's returns
track the benchmark market index's returns. On the other hand, the
Treynor Ratio is a metric used to evaluate an investment portfolio's
risk-adjusted returns.

They have been incorporated in the same PR to simplify the merging, and
because such parameters both refer to the market index.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Frank Milthaler <fmilthaler@users.noreply.github.com>